requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=1.2.0
langchain-openai>=0.3.0
langchain-core>=1.2.0
langgraph>=1.1.0
openai>=1.50.0
pandas>=2.2.0
pandas-ta>=0.3.14b
numpy>=1.26.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
redis>=5.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
yfinance>=0.2.40
numba>=0.59.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
